Support patching libbluetooth.so on Android 10.
When an Android 10 device is detected and the zip file name includes
_btfix, the system's libbluetooth.so library will be patched to allow
Bluetooth pairings to be retained across reboots.

This offers a permanent replacement approach to the run-time masking
approach taken by the libsecure_storage Magisk module.

Tested and verified on S10 series BSKO firmware.

Thank you to Arthur Trouillot for finding the byte sequence to patch.