Device support detection now allows ANY Samsung device.
Detection of a supported device has become too complicated and creates a
maintenance burden.

Adding support for a new device is often as trivial as adding its model
number, but this takes time and delays device support. We therefore now
assume that any Samsung device is a legitimate target. It's up to the
user not to run the multidisabler on a device that is too old.